主机关联查询
