+++ src/common/util/lib.go
@@ -16,87 +16,9 @@ import (
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"net/http"
-	"reflect"
-	"strings"
 
 	restful "github.com/emicklei/go-restful"
 )
-
-func InArray(obj interface{}, target interface{}) bool {
-	targetValue := reflect.ValueOf(target)
-	switch reflect.TypeOf(target).Kind() {
-	case reflect.Slice, reflect.Array:
-		for i := 0; i < targetValue.Len(); i++ {
-			if targetValue.Index(i).Interface() == obj {
-				return true
-			}
-		}
-	case reflect.Map:
-		if targetValue.MapIndex(reflect.ValueOf(obj)).IsValid() {
-			return true
-		}
-	}
-	return false
-}
-
-func ArrayUnique(a interface{}) (ret []interface{}) {
-	va := reflect.ValueOf(a)
-	for i := 0; i < va.Len(); i++ {
-		v := va.Index(i).Interface()
-		if !InArray(v, ret) {
-			ret = append(ret, v)
-		}
-	}
-	return ret
-}
-
-//StrArrayUnique get unique string array
-func StrArrayUnique(a []string) (ret []string) {
-	length := len(a)
-	for i := 0; i < length; i++ {
-		if !Contains(ret, a[i]) {
-			ret = append(ret, a[i])
-		}
-	}
-	return ret
-}
-
-//IntArrayUnique get unique int array
-func IntArrayUnique(a []int) (ret []int) {
-	length := len(a)
-	for i := 0; i < length; i++ {
-		if !ContainsInt(ret, a[i]) {
-			ret = append(ret, a[i])
-		}
-	}
-	return ret
-}
-
-func RemoveDuplicatesAndEmpty(slice []string) (ret []string) {
-	for _, a := range slice {
-		if strings.TrimSpace(a) != "" && !Contains(ret, a) {
-			ret = append(ret, a)
-		}
-	}
-	return
-}
-
-func StrArrDiff(slice1 []string, slice2 []string) []string {
-	var diffStr []string
-	for _, i := range slice1 {
-		isIn := false
-		for _, j := range slice2 {
-			if i == j {
-				isIn = true
-				break
-			}
-		}
-		if !isIn {
-			diffStr = append(diffStr, i)
-		}
-	}
-	return diffStr
-}
 
 // GetActionLanguage returns language form hender
 func GetActionLanguage(req *restful.Request) string {
